djangobackend/requirements.txt
django
djangorestframework 
django-cors-headers
Couchdb
